commerce/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
channels
channels_redis
debugpy
django-filter
daphne
psycopg2-binary